remove stages and slurm.py
